+++ Changelog.md
@@ -14,6 +14,7 @@ v2.33.0 (not published yet)
 ====================
 
 - UI: Enhance "HDR metadata extraction"
+- UI: Optimize "Output Highlighting"
 - UI: Fix crash on "HDR metadata extraction" on some systems
 - UI: Fix language recognition of subtitles from paths ([#1174](/../../issues/1174))
 - UI: Let "Frame Mismatch" start at an earler stage